Tweak emacs startup
It was a little rough bringing emacs up on my devserver with the
git version of the .emacs file; it's mostly because there were a bunch
of packages that I wanted installed that I hadn't installed
locally. But getting that fixed means moving the package install stuff
up above the stuff that needed the packages, so do that.